GraphBLAS: Silence compiler warnings about values outside `GrB_Field`

+++ GraphBLAS/Source/GxB_Context_get2.c
@@ -150,7 +150,7 @@ GrB_Info GxB_Context_get_INT
     // get the field
     //--------------------------------------------------------------------------
 
-    switch (field)
+    switch ((int) field)
     {
 
         case GxB_CONTEXT_NTHREADS :         // same as GxB_NTHREADS
